Depository logic optimization

+++ contracts/GenericBondCalculator.sol
@@ -24,20 +24,21 @@ contract GenericBondCalculator {
 
     /// @dev Calculates the amount of OLAS tokens based on the bonding calculator mechanism.
     /// @notice Currently there is only one implementation of a bond calculation mechanism based on the UniswapV2 LP.
+    /// @notice IDF has a 10^18 multiplier and priceLP has the same as well, so the result must be divided by 10^36.
     /// @param tokenAmount LP token amount.
     /// @param priceLP LP token price.
     /// @return amountOLAS Resulting amount of OLAS tokens.
-    function calculatePayoutOLAS(uint224 tokenAmount, uint256 priceLP) external view
-        returns (uint96 amountOLAS)
+    function calculatePayoutOLAS(uint256 tokenAmount, uint256 priceLP) external view
+        returns (uint256 amountOLAS)
     {
         // The result is divided by additional 1e18, since it was multiplied by in the current LP price calculation
         uint256 amountDF = (ITokenomics(tokenomics).getLastIDF() * priceLP * tokenAmount) / 1e36;
-        amountOLAS = uint96(amountDF);
+        amountOLAS = amountDF;
     }
 
     /// @dev Gets current reserves of OLAS / totalSupply of LP tokens.
     /// @param token Token address.
-    /// @return priceLP Resulting reserveX/totalSupply ratio with 18 decimals.
+    /// @return priceLP Resulting reserveX / totalSupply ratio with 18 decimals.
     function getCurrentPriceLP(address token) external view returns (uint256 priceLP)
     {
         IUniswapV2Pair pair = IUniswapV2Pair(address(token));
@@ -51,7 +52,7 @@ contract GenericBondCalculator {
             (reserve0, reserve1, ) = pair.getReserves();
             // token0 != olas && token1 != olas, this should never happen
             if (token0 == olas || token1 == olas) {
-                // If OLAS is in token0, assign its reserves to reserve1, otherwise the reserve1 is already correct
+                // If OLAS is in token0, assign its reserve to reserve1, otherwise the reserve1 is already correct
                 if (token0 == olas) {
                     reserve1 = reserve0;
                 }
